// app/interfaces/Top.ts
export default interface TopInterface {
    value: {
        nav: { id: any; className: any; };
        ul: { id: any; className: any; };
        lis: { className: any; key: any; content: any; }[];
        as: { href: any; target: any; content: any; }[];
        a: { href: any; target: any; content: any; };
        img: { id: any; className: any; src: any; alt: any; };
    }
}