// app/interfaces/AboutMe.ts
export default interface AboutMeInterface {
    value: {
        section: { id: any; className: any; };
        img: { id: any; className: any; src: any; alt: any; };
        div: { id: any; className: any; };
        h1: { id: any; className: string; content: any; };
        zySpan: { id: any; className: string; content: any; };
        zzSpan: { id: any; className: string; content: any; };
        aaSpan: { id: any; className: string; content: any; };
        abSpan: { id: any; className: string; content: any; };
        acSpan: { id: any; className: string; content: any; };
        ul: { id: any; className: any; };
        lis: { className: any; key: any; content: any; }[];
        button: { id: any; className: string; content: any; };
        openWindow: { url: any; target: any; };
    }
}