// app/interfaces/WhatIOffer.ts
export default interface WhatIOfferInterface {
    value: {
        section: { id: any; className: any; };
        h1: { id: any; className: string; content: any; };
        aaSpan: { id: any; className: string; content: any; };
        abSpan: { id: any; className: string; content: any; };
        h2: { id: any; className: string; content: any; };
        ul: { id: any; className: any; };
        lis: { className: any; key: any; content: any; }[];
    }
}