// app/interfaces/MyPortfolio.ts
export default interface MyPortfolioInterface {
    value: {
        section: { id: any; className: any; };
        h1: { id: any; className: string; content: any; };
        aaSpan: { id: any; className: string; content: any; };
        abSpan: { id: any; className: string; content: any; };
        aaH2: { id: any; className: string; content: any; };
        acSpan: { id: any; className: string; content: any; };
        aaA: { href: string; target: any; content: any; };
        abH2: { id: any; className: string; content: any; };
        adSpan: { id: any; className: string; content: any; };
        abA: { href: string; target: any; content: any; };
    }
}