// app/interfaces/LetsConnect.ts
export default interface LetsConnectInterface {
    value: {
        section: { id: any; className: any; };
        h1: { id: any; className: string; content: any; };
        aaSpan: { id: any; className: string; content: any; };
        abSpan: { id: any; className: string; content: any; };
        h2: { id: any; className: string; content: any; };
        a: { href: string; target: any; content: any; };
        button: { id: any; className: string; content: any; };
        openWindow: { url: any; target: any; };
    }
}