// app/interfaces/Bottom.ts
export default interface BottomInterface {
    value: {
        footer: { id: any; className: any; };
        ul: { id: any; className: any; };
        lis: { className: any; key: any; content: any; }[];
        as: { href: any; target: any; content: any; }[];
        imgs: { id: any; className: any; src: any; alt: any; }[];
        divs: { id: any; className: any; content: any; }[];
        a: { href: any; target: any; content: any; };
        img: { id: any; className: any; src: any; alt: any; };
    }
}